Fix Wyre header
