+++ src/navigation/config.js
@@ -1,8 +1,11 @@
 import React from 'react';
 import { Keyboard } from 'react-native';
 import styled from 'styled-components';
+import BackButton from '../components/header/BackButton';
 import { Icon } from '../components/icons';
 import { SheetHandleFixedToTopHeight } from '../components/sheet';
+import { Text } from '../components/text';
+import { useTheme } from '../context/ThemeContext';
 import colors from '../context/currentColors';
 import { onWillPop } from './Navigation';
 import WalletBackupStepTypes from '@rainbow-me/helpers/walletBackupStepTypes';
@@ -227,16 +230,42 @@ const headerConfigOptions = {
   },
 };
 
-export const wyreWebviewOptions = {
+const EmptyButtonPlaceholder = styled.View`
+  flex: 1;
+`;
+
+const SettingsTitle = ({ children }) => {
+  const { colors } = useTheme();
+
+  return (
+    <Text
+      align="center"
+      color={colors.dark}
+      letterSpacing="roundedMedium"
+      size="large"
+      weight="bold"
+    >
+      {children}
+    </Text>
+  );
+};
+
+export const wyreWebviewOptions = colors => ({
   ...headerConfigOptions,
+  // eslint-disable-next-line react/display-name
+  headerLeft: props => <BackButton {...props} textChevron />,
   headerStatusBarHeight: 24,
   headerStyle: {
-    backgroundColor: colors.themedColors.white,
+    backgroundColor: colors.white,
     elevation: 24,
     shadowColor: 'transparent',
   },
+  headerTitleStyle: {
+    ...headerConfigOptions.headerTitleStyle,
+    color: colors.dark,
+  },
   title: 'Add Cash',
-};
+});
 
 export const settingsOptions = colors => ({
   ...headerConfigOptions,
@@ -256,6 +285,10 @@ export const settingsOptions = colors => ({
     height: 49,
     shadowColor: 'transparent',
   },
+  headerTitleStyle: {
+    ...headerConfigOptions.headerTitleStyle,
+    color: colors.dark,
+  },
   transitionSpec: {
     close: {
       animation: 'spring',
@@ -266,4 +299,12 @@ export const settingsOptions = colors => ({
       config: transitionConfig,
     },
   },
+  ...(android && {
+    // eslint-disable-next-line react/display-name
+    headerLeft: props => <BackButton {...props} textChevron />,
+    // eslint-disable-next-line react/display-name
+    headerRight: () => <EmptyButtonPlaceholder />,
+    // eslint-disable-next-line react/display-name
+    headerTitle: props => <SettingsTitle {...props} />,
+  }),
 });
